requirements.txt
markdown==3.5.2
python-frontmatter==1.1.0
PyYAML==6.0.1 